requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
aiohttp==3.8.4
python-dotenv==1.0.0
loguru==0.7.0
chromadb==0.3.28
sentence-transformers==2.2.2
numpy==1.26.4
pillow==10.0.1
pytesseract==0.3.10
python-magic-bin==0.4.14
pdfplumber==0.9.0
python-telegram-bot==20.6
apscheduler==3.10.1
simhash==2.4.1
sn-scrape==0.4.4
gTTS==2.4.1
requests==2.31.0
boto3==1.28.32
